stat: move unit tests to stat.rs instead of test_stat.rs
